backend/conversation_bot/requirements.txt
openai
firebase-admin
functions-framework
requests
google-auth
tiktoken